+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - TBD
+
+### Added
+- Added `CUSTOM_HEADER_FILE` environment variable to allow inserting custom HTML into the page header.
+- Added `CUSTOM_FOOTER_FILE` environment variable to allow inserting custom HTML into the page footer.
+- Updated `docker-compose.yml` with examples for using `CUSTOM_HEADER_FILE` and `CUSTOM_FOOTER_FILE` with mounts.
+- Added a screenshot to the documentation index.
+
 ## 1.0.0 - 2026-04-27
 
 - Initial release of the Philterd Redaction Policy Editor.